+++ app/src/main/kotlin/com/simplemobiletools/gallery/pro/activities/MainActivity.kt
@@ -459,7 +459,7 @@ class MainActivity : SimpleActivity(), DirectoryOperationsListener {
     }
 
     private fun changeViewType() {
-        ChangeViewTypeDialog(this) {
+        ChangeViewTypeDialog(this, true) {
             invalidateOptionsMenu()
             setupLayoutManager()
             directories_grid.adapter = null
@@ -787,10 +787,10 @@ class MainActivity : SimpleActivity(), DirectoryOperationsListener {
         runOnUiThread {
             checkPlaceholderVisibility(dirs)
 
-            val allowHorizontalScroll = config.scrollHorizontally && config.viewTypeFiles == VIEW_TYPE_GRID
+            val allowHorizontalScroll = config.scrollHorizontally && config.viewTypeFolders == VIEW_TYPE_GRID
             directories_vertical_fastscroller.beVisibleIf(directories_grid.isVisible() && !allowHorizontalScroll)
             directories_horizontal_fastscroller.beVisibleIf(directories_grid.isVisible() && allowHorizontalScroll)
-            setupAdapter(dirs)
+            setupAdapter(dirs.clone() as ArrayList<Directory>)
         }
 
         // cached folders have been loaded, recheck folders one by one starting with the first displayed